bbox uses f32

+++ src/geo/bbox.rs
@@ -2,14 +2,14 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Serialize, Deserialize)]
 pub struct GeoBBox {
-	x_min: f64,
-	x_max: f64,
-	y_min: f64,
-	y_max: f64,
+	x_min: f32,
+	x_max: f32,
+	y_min: f32,
+	y_max: f32,
 }
 
 impl GeoBBox {
-	pub fn new(x_min: f64, x_max: f64, y_min: f64, y_max: f64) -> Self {
+	pub fn new(x_min: f32, x_max: f32, y_min: f32, y_max: f32) -> Self {
 		GeoBBox {
 			x_min,
 			x_max,
@@ -19,23 +19,23 @@ impl GeoBBox {
 	}
 	pub fn new_empty() -> Self {
 		GeoBBox {
-			x_min: f64::MAX,
-			x_max: f64::MIN,
-			y_min: f64::MAX,
-			y_max: f64::MIN,
+			x_min: f32::MAX,
+			x_max: f32::MIN,
+			y_min: f32::MAX,
+			y_max: f32::MIN,
 		}
 	}
 	fn from_vec(v1: &[f64]) -> Self {
 		GeoBBox {
-			x_min: v1[0],
-			x_max: v1[0],
-			y_min: v1[1],
-			y_max: v1[1],
+			x_min: v1[0] as f32,
+			x_max: v1[0] as f32,
+			y_min: v1[1] as f32,
+			y_max: v1[1] as f32,
 		}
 	}
 	fn from_vec2(v2: &[Vec<f64>]) -> Self {
 		let mut bbox = GeoBBox::new_empty();
-		v2.iter().for_each(|v1| bbox.include_point(v1[0], v1[1]));
+		v2.iter().for_each(|v1| bbox.include_point(v1[0] as f32, v1[1] as f32));
 		bbox
 	}
 	fn from_vec3(v3: &[Vec<Vec<f64>>]) -> Self {
@@ -64,7 +64,7 @@ impl GeoBBox {
 			}
 		}
 	}
-	fn include_point(&mut self, x: f64, y: f64) {
+	fn include_point(&mut self, x: f32, y: f32) {
 		if self.x_min > x {
 			self.x_min = x
 		}
@@ -110,10 +110,10 @@ impl GeoBBox {
 		}
 		true
 	}
-	pub fn sum_x(&self) -> f64 {
+	pub fn sum_x(&self) -> f32 {
 		self.x_min + self.x_max
 	}
-	pub fn sum_y(&self) -> f64 {
+	pub fn sum_y(&self) -> f32 {
 		self.y_min + self.y_max
 	}
 }
